Remove python-libaas as automatic requirement
It has been added as an optional dependency to support Canonical MaaS [1].

python-libmaas is not in Fedora so it can not be added as optional
dependency. Watcher code is ready to work without libmaas installed so
we'll keep it without MaaS support until some maintainer shows up to
maintain it.

[1] https://opendev.org/openstack/watcher/commit/c95ce4ec17aa7844f64e97d67cf66c017d656c47

Change-Id: I7189d03ee2449647b3684807cc326b3a3bd29c18