Ignore new PyMySQL build requirement
Patch [1] is adding it upstream but CentOS 9 is providing
python3-PyMySQL-0.10.1-6.el9 which is older that the minimal version
specified.

RDO is not running unit tests at build time, so we can safely ignore it.

At some point it may be good to start running unit tests as part of the
build and we may need to ignore MySQLDbEfficacyIndicatorTestCase but
that will be follow up, once upstream is merged.

Change-Id: I7b14cc2f90417190e318a0ce62de7d6e10c53153